+++ substrate-node/pallets/evidence/Cargo.toml
@@ -1,15 +1,21 @@
 [package]
-name = "pallet-erc20"
-version = "0.1.3"
-authors = ["Vz <Vincent.Zou@qq.com>"]
-edition = "2018"
+authors = ['Substrate DevHub <https://github.com/substrate-developer-hub>']
+description = 'FRAME pallet template for defining custom runtime logic.'
+edition = '2018'
+homepage = 'https://substrate.dev'
+license = 'Unlicense'
+name = 'pallet-evidence'
+repository = 'https://github.com/substrate-developer-hub/substrate-node-template/'
+version = '0.1.3'
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [package.metadata.docs.rs]
 targets = ['x86_64-unknown-linux-gnu']
 
-[dependencies]
-codec = {package = "parity-scale-codec", version = "1.3.1", default-features = false, features = ["derive"]}
+[dependencies.codec]
+default-features = false
+features = ['derive']
+package = 'parity-scale-codec'
+version = '1.3.4'
 
 [dependencies.frame-support]
 default-features = false
@@ -23,7 +29,13 @@ git = 'https://github.com/paritytech/substrate.git'
 tag = 'v2.0.0-rc5'
 version = '2.0.0-rc5'
 
-[dev-dependencies.sp-session]
+[dev-dependencies.sp-core]
+default-features = false
+git = 'https://github.com/paritytech/substrate.git'
+tag = 'v2.0.0-rc5'
+version = '2.0.0-rc5'
+
+[dev-dependencies.sp-io]
 default-features = false
 git = 'https://github.com/paritytech/substrate.git'
 tag = 'v2.0.0-rc5'
@@ -35,13 +47,7 @@ git = 'https://github.com/paritytech/substrate.git'
 tag = 'v2.0.0-rc5'
 version = '2.0.0-rc5'
 
-[dev-dependencies.sp-core]
-default-features = false
-git = 'https://github.com/paritytech/substrate.git'
-tag = 'v2.0.0-rc5'
-version = '2.0.0-rc5'
-
-[dependencies.sp-runtime]
+[dev-dependencies.sp-runtime]
 default-features = false
 git = 'https://github.com/paritytech/substrate.git'
 tag = 'v2.0.0-rc5'
@@ -54,5 +60,4 @@ std = [
     'frame-support/std',
     'frame-system/std',
     'sp-std/std',
-    'sp-runtime/std',
 ]
