+++ substrate-node/runtime/Cargo.toml
@@ -159,6 +159,12 @@ package = 'pallet-erc20'
 path = '../pallets/erc20'
 version = '0.1.3'
 
+[dependencies.evidence]
+default-features = false
+package = 'pallet-evidence'
+path = '../pallets/evidence'
+version = '0.1.3'
+
 [dependencies.timestamp]
 default-features = false
 git = 'https://github.com/paritytech/substrate.git'
@@ -207,4 +213,6 @@ std = [
     'timestamp/std',
     'transaction-payment/std',
     'template/std',
+    'erc20/std',
+    'evidence/std',
 ]
